refractor: Move the ExternalLink component to its seperate file
